requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.2.1
django-cache-url==3.4.5
django-crispy-forms==2.4
environs==14.2.0
gunicorn==23.0.0
marshmallow==4.0.0
packaging==25.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
